refactor: refatorado código de submissão do formulário
